style: remove @wordpress/stylelint-config in favour of custom rules
With the release of Stylelint 15.x, numerous rules previously utilized by
@wordpress/stylelint-config have been deprecated. For detailed information on
these changes, please refer to the Stylelint migration guide:
https://stylelint.io/migration-guide/to-15/#deprecated-stylistic-rules.

The WordPress community is actively engaged in discussions to address these
deprecations, which can be followed on the GitHub issue:
WordPress/gutenberg#48142.

In the interim, due to compatibility issues with @stylistic/stylelint-plugin
when extending WordPress rules, we have opted to manually replicate the
WordPress configuration settings, omitting any deprecated rules in the process.